+++ packages/auto_mappr/example/lib/nullable.g.dart
@@ -11,6 +11,11 @@ part of 'nullable.dart';
 // ignore_for_file: require_trailing_commas, unnecessary_parenthesis
 // ignore_for_file: unnecessary_raw_strings
 
+/// {@template auto_mappr:auto_mappr/example/lib/nullable.dart}
+/// Available mappings:
+/// - `UserDto` → `User` -- With default value.
+/// - `NestedDto` → `Nested`.
+/// {@endtemplate}
 class $Mappr {
   Type _typeOf<T>() => T;
 
@@ -18,18 +23,14 @@ class $Mappr {
   ///
   /// When source model is null, returns `whenSourceIsNull` if defined or throws an exception.
   ///
-  /// Available mappings:
-  /// - `UserDto` → `User` -- With default value.
-  /// - `NestedDto` → `Nested`.
+  /// {@macro auto_mappr:auto_mappr/example/lib/nullable.dart}
   TARGET convert<SOURCE, TARGET>(SOURCE? model) => _convert(model)!;
 
   /// Converts from SOURCE to TARGET if such mapping is configured.
   ///
   /// When source model is null, returns `whenSourceIsNull` if defined or null.
   ///
-  /// Available mappings:
-  /// - `UserDto` → `User` -- With default value.
-  /// - `NestedDto` → `Nested`.
+  /// {@macro auto_mappr:auto_mappr/example/lib/nullable.dart}
   TARGET? tryConvert<SOURCE, TARGET>(SOURCE? model) => _convert(
         model,
         canReturnNull: true,
@@ -39,9 +40,7 @@ class $Mappr {
   ///
   /// When an item in the source iterable is null, uses `whenSourceIsNull` if defined or throws an exception.
   ///
-  /// Available items mappings:
-  /// - `UserDto` → `User` -- With default value.
-  /// - `NestedDto` → `Nested`.
+  /// {@macro auto_mappr:auto_mappr/example/lib/nullable.dart}
   Iterable<TARGET> convertIterable<SOURCE, TARGET>(Iterable<SOURCE?> model) =>
       model.map<TARGET>((item) => _convert(item)!);
 
@@ -49,9 +48,7 @@ class $Mappr {
   ///
   /// When an item in the source iterable is null, uses `whenSourceIsNull` if defined or null
   ///
-  /// Available items mappings:
-  /// - `UserDto` → `User` -- With default value.
-  /// - `NestedDto` → `Nested`.
+  /// {@macro auto_mappr:auto_mappr/example/lib/nullable.dart}
   Iterable<TARGET?> tryConvertIterable<SOURCE, TARGET>(
           Iterable<SOURCE?> model) =>
       model.map<TARGET?>((item) => _convert(item, canReturnNull: true));
@@ -60,42 +57,33 @@ class $Mappr {
   ///
   /// When an item in the source iterable is null, uses `whenSourceIsNull` if defined or throws an exception.
   ///
-  /// Available items mappings:
-  /// - `UserDto` → `User` -- With default value.
-  /// - `NestedDto` → `Nested`.
+  /// {@macro auto_mappr:auto_mappr/example/lib/nullable.dart}
   List<TARGET> convertList<SOURCE, TARGET>(Iterable<SOURCE?> model) =>
-      model.map<TARGET>((item) => _convert(item)!).toList();
+      convertIterable<SOURCE, TARGET>(model).toList();
 
   /// For iterable items, converts from SOURCE to TARGET if such mapping is configured, into List.
   ///
   /// When an item in the source iterable is null, uses `whenSourceIsNull` if defined or null
   ///
-  /// Available items mappings:
-  /// - `UserDto` → `User` -- With default value.
-  /// - `NestedDto` → `Nested`.
-  List<TARGET?> tryConvertList<SOURCE, TARGET>(Iterable<SOURCE?> model) => model
-      .map<TARGET?>((item) => _convert(item, canReturnNull: true))
-      .toList();
+  /// {@macro auto_mappr:auto_mappr/example/lib/nullable.dart}
+  List<TARGET?> tryConvertList<SOURCE, TARGET>(Iterable<SOURCE?> model) =>
+      convertIterable<SOURCE, TARGET>(model).toList();
 
   /// For iterable items, converts from SOURCE to TARGET if such mapping is configured, into Set.
   ///
   /// When an item in the source iterable is null, uses `whenSourceIsNull` if defined or throws an exception.
   ///
-  /// Available items mappings:
-  /// - `UserDto` → `User` -- With default value.
-  /// - `NestedDto` → `Nested`.
+  /// {@macro auto_mappr:auto_mappr/example/lib/nullable.dart}
   Set<TARGET> convertSet<SOURCE, TARGET>(Iterable<SOURCE?> model) =>
-      model.map<TARGET>((item) => _convert(item)!).toSet();
+      convertIterable<SOURCE, TARGET>(model).toSet();
 
   /// For iterable items, converts from SOURCE to TARGET if such mapping is configured, into Set.
   ///
   /// When an item in the source iterable is null, uses `whenSourceIsNull` if defined or null
   ///
-  /// Available items mappings:
-  /// - `UserDto` → `User` -- With default value.
-  /// - `NestedDto` → `Nested`.
+  /// {@macro auto_mappr:auto_mappr/example/lib/nullable.dart}
   Set<TARGET?> tryConvertSet<SOURCE, TARGET>(Iterable<SOURCE?> model) =>
-      model.map<TARGET?>((item) => _convert(item, canReturnNull: true)).toSet();
+      convertIterable<SOURCE, TARGET>(model).toSet();
   TARGET? _convert<SOURCE, TARGET>(
     SOURCE? model, {
     bool canReturnNull = false,